Refactor tenpy.linalg submodule structure
Remove tenpy.linalg.symmetries folder. Move modules
tenpy.linalg.symmetries.groups -> tenpy.linalg.groups
tenpy.linalg.symmetries.spaces -> tenpy.linalg.spaces

Hide substructure of tenpy.linalg.backends, i.e. expose everything
in tenpy.linalg.backends.
