+++ tenpy/linalg/backends/__init__.py
@@ -1,8 +1,17 @@
 # Copyright 2023-2023 TeNPy Developers, GNU GPLv3
-from . import abstract_backend, abelian, backend_factory, no_symmetry, nonabelian, numpy, torch
-from .abstract_backend import AbstractBackend
-from .numpy import NoSymmetryNumpyBackend, AbelianNumpyBackend, NonabelianNumpyBackend
-from .torch import NoSymmetryTorchBackend, AbelianTorchBackend, NonabelianTorchBackend
-from .backend_factory import get_backend
-__all__ = ['backend_factory', 'abstract_backend', 'no_symmetry', 'abelian', 'nonabelian', 'numpy', 
-           'torch']
+
+# hide the folder-structure and expose everyhting as if everything was implemented
+# directly in tenpy.linalg.backends
+from . import abelian, abstract_backend, backend_factory, no_symmetry, nonabelian, numpy, torch
+from .abelian import *
+from .abstract_backend import *
+from .array_api import *
+from .backend_factory import *
+from .no_symmetry import *
+from .nonabelian import *
+from .numpy import *
+from .torch import *
+__all__ = [
+    *abelian.__all__, *abstract_backend.__all__, *backend_factory.__all__, *no_symmetry.__all__,
+    *nonabelian.__all__, *numpy.__all__, *torch.__all__
+]
