+++ tenpy/linalg/backends/numpy.py
@@ -10,7 +10,7 @@ from .abstract_backend import AbstractBlockBackend, Block, Data, Dtype
 from .no_symmetry import AbstractNoSymmetryBackend
 from .nonabelian import AbstractNonabelianBackend
 from ..misc import inverse_permutation
-from ..symmetries.spaces import VectorSpace
+from ..spaces import VectorSpace
 
 __all__ = ['NumpyBlockBackend', 'NoSymmetryNumpyBackend', 'AbelianNumpyBackend', 'NonabelianNumpyBackend']
