--- tenpy/linalg/symmetries/__init__.py
@@ -1,4 +1,0 @@
-# Copyright 2023-2023 TeNPy Developers, GNU GPLv3
-from . import groups, spaces
-
-__all__ = ['groups', 'spaces']
