+++ tests/linalg/test_spaces.py
@@ -3,7 +3,7 @@ import pytest
 import numpy as np
 from numpy.testing import assert_array_equal
 
-from tenpy.linalg.symmetries import spaces, groups
+from tenpy.linalg import spaces, groups
 from tenpy.linalg.backends import abelian
