+++ tests/linalg/test_tensors.py
@@ -11,7 +11,7 @@ from tenpy.linalg.backends.no_symmetry import AbstractNoSymmetryBackend
 from tenpy.linalg.backends.abelian import AbstractAbelianBackend
 from tenpy.linalg.backends.torch import TorchBlockBackend
 from tenpy.linalg.backends.numpy import NumpyBlockBackend, NoSymmetryNumpyBackend
-from tenpy.linalg.symmetries.spaces import VectorSpace, ProductSpace, _fuse_spaces
+from tenpy.linalg.spaces import VectorSpace, ProductSpace, _fuse_spaces
 
 
 
@@ -291,7 +291,7 @@ def test_ChargedTensor_tofrom_flat_block_single_sector(vector_space_rng, symmetr
 
 
 def test_from_block(block_backend):
-    from tenpy.linalg.symmetries.groups import z4_symmetry
+    from tenpy.linalg.groups import z4_symmetry
     from tenpy.linalg.backends.backend_factory import get_backend
     backend = get_backend(symmetry=z4_symmetry, block_backend=block_backend)
